requirements.txt
Django==3.0.1
Celery==4.4.0
redis==3.3.11
psycopg2==2.8.4
requests==2.22.0
django-environ==0.4.5
graphene-django==2.7.1
django-cors-headers==3.2.0
twitchio==1.0.0
channels==2.4.0
channels-redis==2.4.1
dj-database-url==0.5.0